# [Enhancement] Use RangeDirectMapping to optimize hash join