… be/src/exec/join/join_hash_map.tpp
… be/src/exec/join/join_hash_map_method.hpp
… be/src/exec/join/join_hash_table_descriptor.h
… be/src/exec/join/join_type_traits.h
… be/src/util/runtime_profile.h
… be/test/exec/join_hash_map_test.cpp
… fe/fe-core/src/main/java/com/starrocks/qe/SessionVariable.java
… test/sql/test_join/R/test_join_range_direct_mapping
… test/sql/test_join/T/test_join_range_direct_mapping