## Why I'm doing:


### Adding Implementations for Three RANGE_DIRECT_MAPPING Hash Map Methods
Add three hash map method implementations: `RANGE_DIRECT_MAPPING`, `RANGE_DIRECT_MAPPING_SET`, `DENSE_RANGE_DIRECT_MAPPING`.


#### Implementation Details
Compared to the default BUCKET_CHAINED:
- The similarity lies in the bucket-chained structure formed by first and next.
- The difference is that for a given key, `BUCKET_CHAINED` uses a hash function to map to a position in first, while RANGE_DIRECT_MAPPING uses `key-min_value` and its variants to directly map to first's position.
    - `RANGE_DIRECT_MAPPING`: Uses `key-min_value` as the index for first.
    - `RANGE_DIRECT_MAPPING_SET`: For `LEFT SEMI/ANTI JOIN` (without other conjuncts), replaces `first` and `next` with a `key_bitset`, where each `key-min_value` position is represented by a single bit to indicate existence.
    - `DENSE_RANGE_DIRECT_MAPPING`: Compresses `max_value-min_value` `first` positions down to `row_count` positions by discarding empty `first` slots to save memory. Each dense group stores information for 32 first positions using 64 bits (start_index: 32 bits, bitset: 32 bits). Thus, each `value_interval` position is represented with 2 bits (64 bits / 32 = 2 bits).
       - `start_index`: Indicates the starting position of this dense group within the dense first array.
       - `bitset`: Uses 32 bits to mark which of the 32 positions in this dense group are non-empty.

<br>

**BUCKET_CHAINED**

```
 build keys                       first       next
                                  ┌───┐       ┌───┐
                                  │   │       │   │◄───┐
                                  │   │       │   │◄┐  │
                                  ├───┤       ├───┤ │  │
                                  │   ├─┐     │   │ │  │
                         ┌───────►│   │ │     │   │ │  │
                         │        ├───┤ │     ├───┤ │  │
              ┌────────┐ │        │   │ │     │   ├─┘  │
  ┌──────┐    │        │ │        │   │ │     │   │◄─┐ │
  │ key  ├───►│  Hash  ├─┘        ├───┤ │     ├───┤  │ │
  └──────┘    │        │          │   │ │     │   │  │ │
              └────────┘          │   │ │     │   │  │ │
                                  ├───┤ │     ├───┤  │ │
                                  │   │ └────►│   │  │ │
                                  │   │       │   ├──┘ │
                                  ├───┤       ├───┤    │
                                  │   │       │   │    │
                                  │   │       │   │    │
                                  ├───┤       ├───┤    │
                                  │   │       │   │    │
                                  │   ├──────►│   ├────┘
                                  └───┘       └───┘
```

<br>

**RANGE_DIRECT_MAPPING**

```
 build keys               first       next
                          ┌───┐       ┌───┐
                          │   │       │   │◄───┐
                          │   │       │   │◄┐  │
                          ├───┤       ├───┤ │  │
          key-min_value   │   ├─┐     │   │ │  │
                 ┌───────►│   │ │     │   │ │  │
 ┌──────┐        │        ├───┤ │     ├───┤ │  │
 │ key  ├────────┘        │   │ │     │   ├─┘  │
 └──────┘                 │   │ │     │   │◄─┐ │
                          ├───┤ │     ├───┤  │ │
                          │   │ │     │   │  │ │
                          │   │ │     │   │  │ │
                          ├───┤ │     ├───┤  │ │
                          │   │ └────►│   │  │ │
                          │   │       │   ├──┘ │
                          ├───┤       ├───┤    │
                          │   │       │   │    │
                          │   │       │   │    │
                          ├───┤       ├───┤    │
                          │   │       │   │    │
                          │   ├──────►│   ├────┘
                          └───┘       └───┘
```


<br>

**DENSE_RANGE_DIRECT_MAPPING**

```
 build keys              dense groups       first       next
                         ┌───────────┐      ┌───┐       ┌───┐
                         │start_index┼─┐    │   │       │ 0 │◄───┐
                         │bitset     │ │    │   │       │   │◄┐  │
                         ├───────────┤ │    ├───┤       ├───┤ │  │
                         │           │ │    │   │─┐     │   │ │  │
                         │           │ └───►│   │ │     │   │ │  │
                         ├───────────┤      ├───┤ │     ├───┤ │  │
                         │           │      │   │ │     │   │─┘  │
  ┌──────┐key-min_value  │           │      │   │ │     │   │◄─┐ │
  │ key  ├─────────────► ├───────────┤      ├───┤ │     ├───┤  │ │
  └──────┘               │           │      │   │ │     │   │  │ │
                         │           │      │   │ │     │   │  │ │
                         ├───────────┤      ├───┤ │     ├───┤  │ │
                         │           │      │   │ └────►│   │  │ │
                         │           │      │   │       │   │──┘ │
                         ├───────────┤      ├───┤       ├───┤    │
                         │           │      │   │       │   │    │
                         │           │      │   │       │   │    │
                         ├───────────┤      ├───┤       ├───┤    │
                         │           │      │   │       │   │    │
                         │           │      │   │──────►│   │────┘
                         └───────────┘      └───┘       └───┘
```

#### Usage Scenarios
General principle: 
- Key Column is `INT` or `BIGINT`.
- Use when it doesn’t increase memory compared to BUCKET_CHAINED, that is, `max_value - min_value + 1 <= bucket_size`.

For details, refer to `_try_use_range_direct_mapping` in the file `join_hash_map.cpp`.

#### Benefits
- Eliminates the need for hashing.
- All entries in a single bucket's linked list have identical values, meaning:
    - During probing, there’s no need to compare the probe key against the build key.
    - When builder keys are unique, accessing the `table_items.next` array becomes unnecessary.
- For LEFT SEMI/ANTI JOIN, a bitset suffices without needing the bucket-chained structure of first and next.

### Adding HashMapType Field to Profile
Add a HashMapType field in the profile, formatted as `KeyConstructorType-MapMethodType-KeyLogicalType:<NumPartitions>,...`  
For example:
- HashMapType: ONE_KEY-RANGE_DIRECT_MAPPING-INT:1
- HashMapType: SERIALIZED_KEY-BUCKET_CHAINED-VARCHAR:16


### Test
Environment
- 4 BE (each 16 vCPUs, 64 GB Memory)

Dadataset
- ssb 1T

<br>

**RANGE_DIRECT_MAPPING_SET**

```sql
SELECT SUM(lo_revenue)  FROM lineorder  LEFT SEMI JOIN customer ON lo_custkey=c_custkey and c_region='ASIA';
```
- Before: 11.482s
- After: 5.685s

<br>

**DENSE_RANGE_DIRECT_MAPPING**

```sql
SELECT SUM(lo_revenue), SUM(c_custkey)  FROM lineorder  LEFT JOIN customer ON lo_custkey=c_custkey WHERE c_region='ASIA';
```
- Before: 8.361s
- After: 6.241s


<br>

**RANGE_DIRECT_MAPPING**

```sql
SELECT SUM(lo_revenue), SUM(s_suppkey)  FROM lineorder  LEFT JOIN supplier ON lo_suppkey=s_suppkey;
```
- Before: 8.980s
- After: 3.255s

<br>

**SSB standard queries:**

  | Before | After | Before/After
-- | -- | -- | --
SUM | 47510 | 44517 | 1.07
Q01 | 1802 | 1763 | 1.02
Q02 | 1450 | 1430 | 1.01
Q03 | 1319 | 1351 | 0.98
Q04 | 4753 | 4700 | 1.01
Q05 | 3431 | 3369 | 1.02
Q06 | 3226 | 3314 | 0.97
Q07 | 9028 | 7754 | 1.16
Q08 | 3322 | 3343 | 0.99
Q09 | 3166 | 3128 | 1.01
Q10 | 1673 | 1664 | 1.01
Q11 | 7921 | 6500 | 1.22
Q12 | 3556 | 3410 | 1.04
Q13 | 2863 | 2791 | 1.03


## What I'm doing:


Fixes #issue

## What type of PR is this:

- [ ] BugFix
- [ ] Feature
- [x] Enhancement
- [ ] Refactor
- [ ] UT
- [ ] Doc
- [ ] Tool

Does this PR entail a change in behavior?

- [ ] Yes, this PR will result in a change in behavior.
- [x] No, this PR will not result in a change in behavior.

If yes, please specify the type of change:

- [ ] Interface/UI changes: syntax, type conversion, expression evaluation, display information
- [ ] Parameter changes: default values, similar parameters but with different default values
- [ ] Policy changes: use new policy to replace old one, functionality automatically enabled
- [ ] Feature removed
- [ ] Miscellaneous: upgrade & downgrade compatibility, etc.

## Checklist:

- [x] I have added test cases for my bug fix or my new feature
- [ ] This pr needs user documentation (for new or modified features or behaviors)
  - [ ] I have added documentation for my new feature or new function
- [ ] This is a backport pr

## Bugfix cherry-pick branch check:
- [x] I have checked the version labels which the pr will be auto-backported to the target branch
  - [ ] 3.5
  - [ ] 3.4
  - [ ] 3.3